pyproject.toml
[tool.poetry]
name = "twitter-clone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "twitter_clone"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
black = "^22.12.0"
dj-rest-auth = "^2.2.7"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
